# update benchmark & profile to support meta client and commit performance fix for meta protocol