--- a/.github/workflows/benchmarks.yml
+++ b/.github/workflows/benchmarks.yml
@@ -20,7 +20,7 @@ jobs:
     - name: Set up Ruby
       uses: ruby/setup-ruby@v1
       with:
-        ruby-version: 3.2
+        ruby-version: 3.4
         bundler-cache: true # 'bundle install' and cache
     - name: Run Benchmarks
       run: RUBY_YJIT_ENABLE=1 BENCH_TARGET=all bundle exec bin/benchmark