--- a/bin/benchmark
+++ b/bin/benchmark
@@ -2,13 +2,12 @@
 # frozen_string_literal: true
 
 # This helps benchmark current performance of Dalli
-# as well as compare performance of optimizated and non-optimized calls like multi-set vs set
+# as well as compare performance of optimized and non-optimized calls like multi-set vs set
 #
 # run with:
 # bundle exec bin/benchmark
 # RUBY_YJIT_ENABLE=1 BENCH_TARGET=get bundle exec bin/benchmark
 require 'bundler/inline'
-require 'json'
 
 gemfile do
   source 'https://rubygems.org'
@@ -16,14 +15,15 @@ gemfile do
   gem 'logger'
 end
 
-require_relative '../lib/dalli'
+require 'json'
 require 'benchmark/ips'
 require 'monitor'
+require_relative '../lib/dalli'
 
 ##
-# StringSerializer is a serializer that avoids the overhead of Marshal or JSON.
+# NoopSerializer is a serializer that avoids the overhead of Marshal or JSON.
 ##
-class StringSerializer
+class NoopSerializer
   def self.dump(value)
     value
   end
@@ -34,7 +34,7 @@ class StringSerializer
 end
 
 dalli_url = ENV['BENCH_CACHE_URL'] || '127.0.0.1:11211'
-bench_target = ENV['BENCH_TARGET'] || 'set'
+bench_target = ENV['BENCH_TARGET'] || 'all'
 bench_time = (ENV['BENCH_TIME'] || 10).to_i
 bench_warmup = (ENV['BENCH_WARMUP'] || 3).to_i
 bench_payload_size = (ENV['BENCH_PAYLOAD_SIZE'] || 700_000).to_i
@@ -42,9 +42,9 @@ payload = 'B' * bench_payload_size
 TERMINATOR = "\r\n"
 puts "yjit: #{RubyVM::YJIT.enabled?}"
 
-client = Dalli::Client.new(dalli_url, serializer: StringSerializer, compress: false, raw: true)
-meta_client = Dalli::Client.new(dalli_url, protocol: :meta, serializer: StringSerializer, compress: false, raw: true)
-multi_client = Dalli::Client.new('localhost:11211,localhost:11222', serializer: StringSerializer, compress: false,
+client = Dalli::Client.new(dalli_url, serializer: NoopSerializer, compress: false, raw: true)
+meta_client = Dalli::Client.new(dalli_url, protocol: :meta, serializer: NoopSerializer, compress: false, raw: true)
+multi_client = Dalli::Client.new('localhost:11211,localhost:11222', serializer: NoopSerializer, compress: false,
                                                                     raw: true)
 
 # The raw socket implementation is used to benchmark the performance of dalli & the overhead of the various abstractions
@@ -54,7 +54,7 @@ sock.setsockopt(Socket::IPPROTO_TCP, Socket::TCP_NODELAY, true)
 sock.setsockopt(Socket::SOL_SOCKET, Socket::SO_KEEPALIVE, true)
 # Benchmarks didn't see any performance gains from increasing the SO_RCVBUF buffer size
 # sock.setsockopt(Socket::SOL_SOCKET, ::Socket::SO_RCVBUF, 1024 * 1024 * 8)
-# Benchamrks did see an improvement in performance when increasing the SO_SNDBUF buffer size
+# Benchmarks did see an improvement in performance when increasing the SO_SNDBUF buffer size
 # sock.setsockopt(Socket::SOL_SOCKET, Socket::SO_SNDBUF, 1024 * 1024 * 8)
 
 # ensure the clients are all connected and working