--- a/bin/profile
+++ b/bin/profile
@@ -41,6 +41,7 @@ TERMINATOR = "\r\n"
 puts "yjit: #{RubyVM::YJIT.enabled?}"
 
 client = Dalli::Client.new(dalli_url, serializer: StringSerializer, compress: false)
+meta_client = Dalli::Client.new(dalli_url, protocol: :meta, serializer: StringSerializer, compress: false, raw: true)
 
 # The raw socket implementation is used to benchmark the performance of dalli & the overhead of the various abstractions
 # in the library.
@@ -56,6 +57,7 @@ payload = 'B' * bench_payload_size
 dalli_key = 'dalli_key'
 # ensure the clients are all connected and working
 client.set(dalli_key, payload)
+meta_client.set(dalli_key, payload)
 sock.write("set sock_key 0 3600 #{payload.bytesize}\r\n")
 sock.write(payload)
 sock.write(TERMINATOR)
@@ -132,6 +134,14 @@ if %w[all get].include?(bench_target)
     end
   end
 
+  Vernier.profile(out: 'meta_client_get_profile.json') do
+    start_time = Time.now
+    while Time.now - start_time < bench_time
+      result = meta_client.get(dalli_key)
+      raise 'mismatch' unless result == payload
+    end
+  end
+
   Vernier.profile(out: 'socket_get_profile.json') do
     start_time = Time.now
     while Time.now - start_time < bench_time
@@ -148,6 +158,11 @@ if %w[all set].include?(bench_target)
   Vernier.profile(out: 'client_set_profile.json') do
     start_time = Time.now
     client.set(dalli_key, payload, 3600, raw: true) while Time.now - start_time < bench_time
+  end
+
+  Vernier.profile(out: 'meta_client_set_profile.json') do
+    start_time = Time.now
+    meta_client.set(dalli_key, payload, 3600, raw: true) while Time.now - start_time < bench_time
   end
 
   Vernier.profile(out: 'socket_set_profile.json') do
@@ -171,6 +186,14 @@ if %w[all get_multi].include?(bench_target)
     end
   end
 
+  Vernier.profile(out: 'meta_client_get_multi_profile.json') do
+    start_time = Time.now
+    while Time.now - start_time < bench_time
+      result = meta_client.get_multi(pairs.keys)
+      raise 'mismatch' unless result == pairs
+    end
+  end
+
   Vernier.profile(out: 'socket_get_multi_profile.json') do
     start_time = Time.now
     while Time.now - start_time < bench_time
@@ -188,6 +211,15 @@ if %w[all set_multi].include?(bench_target)
     while Time.now - start_time < bench_time
       pairs.each do |key, value|
         client.set(key, value, 3600, raw: true)
+      end
+    end
+  end
+
+  Vernier.profile(out: 'meta_client_set_multi_profile.json') do
+    start_time = Time.now
+    while Time.now - start_time < bench_time
+      pairs.each do |key, value|
+        meta_client.set(key, value, 3600, raw: true)
       end
     end
   end