Ok, with the benchmark and profile we can start to pull over some performance improvements. I updated the benchmark and profile to support the meta protocol client as well as the binary. We haven't been working on the binary, but it is good to have both as we don't want to accidentally cause any regressions with the binary protocol support before we deprecate it.

![dalli_string_profile](https://github.com/user-attachments/assets/490b1470-6ad5-4357-994b-f549a500edf7)
> using vernier profiler on the set command we can see it is spending a lot of time with string manipulation, this can be especially bad with very large values.

The before benchmark, shows that meta set is the slowest 
```
❯❯❯$ BENCH_TARGET=set RUBY_YJIT_ENABLE=1 bundle exec bin/benchmark                                                 <bundler> [main 
yjit: true
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
Warming up --------------------------------------
          client set   730.000 i/100ms
     meta client set   509.000 i/100ms
        raw sock set     1.221k i/100ms
Calculating -------------------------------------
          client set      6.279k (±10.3%) i/s -     62.780k in  10.101866s
     meta client set      4.547k (±13.4%) i/s -     44.792k in  10.012413s
        raw sock set     13.935k (± 2.2%) i/s -    140.415k in  10.082106s

Comparison:
        raw sock set:    13934.8 i/s
          client set:     6278.6 i/s - 2.22x  slower
     meta client set:     4546.7 i/s - 3.06x  slower
```

after benchmark, shows meta is now nearly as fast as raw socket
```
❯❯❯$ BENCH_TARGET=set RUBY_YJIT_ENABLE=1 bundle exec bin/benchmark                                                 <bundler> 
yjit: true
ruby 3.3.0 (2023-12-25 revision 5124f9ac75) +YJIT [arm64-darwin23]
Warming up --------------------------------------
          client set   303.000 i/100ms
     meta client set     1.025k i/100ms
        raw sock set     1.436k i/100ms
Calculating -------------------------------------
          client set      6.573k (±10.8%) i/s -     64.842k in  10.004278s
     meta client set     13.347k (± 6.2%) i/s -    133.250k in  10.024490s
        raw sock set     14.580k (± 5.2%) i/s -    146.472k in  10.073444s

Comparison:
        raw sock set:    14580.3 i/s
     meta client set:    13347.1 i/s - same-ish: difference falls within error
          client set:     6573.4 i/s - 2.22x  slower
```