@danmayer 2025-02-14
> Please keep PRs to individual units of work. Fixing both the bench script and making your change in one pr is a bit confusing.

yeah sorry the primary issue was I realized that the previous PR didn't set clients to meta protocol as it defaults that way in some of our other client setups.